requirements.txt
colpali_engine==0.3.9
fitz==0.0.1.dev2
matplotlib==3.10.1
numpy==2.2.4
pdf2image==1.17.0
Pillow==11.1.0
qdrant_client==1.13.3
Requests==2.32.3
rich==14.0.0
stamina==25.1.0
torch==2.6.0
tqdm==4.67.1
